Float numbers sum

+++ 005_Add_2_Numbers_Joswin.py
@@ -1,6 +1,6 @@
 def add(a, b):
     return a + b
 
-a = int(input("Enter first number : "))
-b = int(input("Enter second number : "))
+a = float(input("Enter first number : "))
+b = float(input("Enter second number : "))
 print(add(a, b))
